requirements.txt
facebook-business>=18.0.0
python-dotenv>=1.0.0
pytz>=2023.3
anthropic>=0.39.0

